requirements.txt
PyPDF2
reportlab
natsorte